requirements.txt
google-genai>=0.8.0
Pillow>=10.0.0
APScheduler>=3.10.4
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
